docs(Dropdown): update Dropdown controls and add playground
